Make git_sha plugin always execute

+++ _plugins/jekyll-labnotebook-plugins/git_sha_modified.rb
@@ -1,13 +1,13 @@
 # Title: git_mod.rb
 # Author: Carl Boettiger, @cboettig
 # Licence: MIT
-# Description: A custom filter to return the date that git has for the page modification time 
-# 
+# Description: A custom filter to return the date that git has for the page modification time
+#
 # REQUIREMENTS: Uses `path` variable set in the base_name.rb plugin
 #
-# Example use: 
+# Example use:
 #
-# 
+#
 
 require 'git'
 require 'json'
@@ -17,7 +17,7 @@ module Jekyll
 
 
     def generate(site)
-  
+
       def niltext(char)
         if(char.nil?)
           "not computed"
@@ -25,38 +25,39 @@ module Jekyll
           char
         end
       end
-     
+
       puts "Getting date modified and sha data from git"
       start = Time.now
       g = Git.open(".")
-      ## FIXME takes over a second per post!  
-      # That's almost 15 minutes at the moment.... Would be much much faster if we could 
+      ## FIXME takes over a second per post!
+      # That's almost 15 minutes at the moment.... Would be much much faster if we could
       # generate a data file and read in as a hash containing the path, sha, and commit time
       # for each post.  Then in looping over posts we could just look up sha and timestamp
-      # from the path in the hash, rather than calling git gem each time.  
-      if(!File.exists?("gitdata.json"))
+      # from the path in the hash, rather than calling git gem each time.
+      ## Make it execute always so that we have fresh data available
+      if true # (!File.exists?("gitdata.json"))
 
-        ## Loop over posts 
+        ## Loop over posts
         site.posts.docs.each do |post|
           begin
             path = '_posts/' + post.basename
             modified = g.log(1).object(path).first.date
             sha = g.log(1).object(path).first.sha
-          rescue Exception => e 
-            modified = Time.now 
-            sha = "not yet committed" 
+          rescue Exception => e
+            modified = Time.now
+            sha = "not yet committed"
           end
           post.data['sha'] = sha
           post.data['modified'] = modified
         end
-      
-        gitdata = Hash[site.posts.docs.collect{|post| 
+
+        gitdata = Hash[site.posts.docs.collect{|post|
           ['_posts/' + post.basename, [post.data['sha'], post.data['modified']]]}]
 
         File.open("gitdata.json", "w") do |f|
               f.write(JSON.pretty_generate(gitdata))
         end
-      end 
+      end
 
       if(File.exists?("gitdata.json"))
         buffer = open("gitdata.json")
@@ -68,15 +69,15 @@ module Jekyll
         end
       end
 
-  ## Loop over pages 
+  ## Loop over pages
       site.pages.each do |page|
         begin
           path = File.join(".", page.dir, page.basename)
           modified = g.log(1).object(path).first.date
           sha = g.log(1).object(path).first.sha
-        rescue Exception => e 
-          modified = Time.now 
-          sha = "not yet committed" 
+        rescue Exception => e
+          modified = Time.now
+          sha = "not yet committed"
         end
         page.data['sha'] = sha
         page.data['modified'] = modified
@@ -88,4 +89,3 @@ module Jekyll
     end
   end
 end
-
